d0.0 Some library tests
