+++ src/pk2_main.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "engine/PDLL.hpp"
+#include <string>
+
+
+std::string PK2_EXPORT pk2_get_version();
+void PK2_EXPORT pk2_main(bool _dev_mode, bool _show_fps, bool test_level, const std::string& test_path);
+bool PK2_EXPORT pk2_convertToSpr2(const std::string& filename_in, const std::string& filename_out);
+bool PK2_EXPORT pk2_setAssetsPath(const std::string& path);
